+++ src/core/types/command-config.ts
@@ -1,6 +1,9 @@
 export interface CommandConfig {
+    module: string;
     id: string;
     icon: string;
     label: string;
-    withLoader: boolean;
+    withLoader?: boolean;
+    loaderLabel?: string;
+    subCommand?: boolean;
 }