+++ src/core/types/command.ts
@@ -1,5 +1,3 @@
-import * as vscode from 'vscode';
-
 export interface Command {
     showOnCommandPalette: boolean;
     execute(): void;