+++ src/extension.ts
@@ -1,10 +1,10 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
 
-import { Views } from './views/_index';
-import { Modules } from './modules/modules.index';
-import { Store } from './store';
+import { Views } from "./views/_index";
+import { Modules } from "./modules/modules.index";
+import { Store } from "./store";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -20,7 +20,7 @@ export function activate(context: vscode.ExtensionContext) {
     Modules.getModulesArray().forEach((module: any) => {
         module.getCommandsArray().forEach((command: any) => {
             if (command.show() && command.showOnCommandPalette) {
-                const commandName = `floorfive-vs-code.${command.getModule()}.${command.getId()}`;
+                const commandName = `floorfive-vs-code.${ command.getModule() }.${ command.getId() }`;
                 let disposable = vscode.commands.registerCommand(commandName, () => {
                     command.execute();
                 });
@@ -35,7 +35,7 @@ export function activate(context: vscode.ExtensionContext) {
 
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
-    console.log(`Congratulations, your extension "floorfive-vs-code" is now active!`);
+    console.log("Congratulations, your extension \"floorfive-vs-code\" is now active!");
 }
 
 // This method is called when your extension is deactivated