+++ src/services/icons.service.ts
@@ -1,6 +1,6 @@
-import * as fs from 'fs';
-import * as vscode from 'vscode';
-import { ICONS } from '../icons';
+import * as fs from "fs";
+import * as vscode from "vscode";
+import { ICONS } from "../icons";
 import { Utilities } from "./utilities.service";
 
 /**
@@ -23,15 +23,15 @@ export class IconsService {
         // Get icons paths
         const iconsPaths: { [key: string]: vscode.Uri } = {};
         for (const icon of this.iconsList) {
-            iconsPaths[icon] = vscode.Uri.joinPath(extensionUri, `assets`, `icons`, `${icon}.svg`);
+            iconsPaths[icon] = vscode.Uri.joinPath(extensionUri, "assets", "icons", `${ icon }.svg`);
         }
 
         let iconsScript = `
 // ==== ICONS ====
 
 var icons = {
-    ${IconsService.iconsList.map(icon => {
-            return `"${icon}": \`{{${Utilities.snakeToCamel(icon)}Icon}}\``;
+    ${ IconsService.iconsList.map(icon => {
+            return `"${ icon }": \`{{${ Utilities.snakeToCamel(icon) }Icon}}\``;
         })
             }
 };
@@ -48,8 +48,8 @@ iconTags.forEach(function(iconTag) {
 
         // Replace template variables with icons actual svg tags, read from files
         for (const key of Object.keys(iconsPaths)) {
-            const icon = fs.readFileSync(iconsPaths[key].fsPath, `utf8`);
-            iconsScript = iconsScript.replace(`{{${Utilities.snakeToCamel(key)}Icon}}`, icon);
+            const icon = fs.readFileSync(iconsPaths[key].fsPath, "utf8");
+            iconsScript = iconsScript.replace(`{{${ Utilities.snakeToCamel(key) }Icon}}`, icon);
         }
 
         return iconsScript;