+++ src/store.ts
@@ -1,13 +1,13 @@
-import * as vscode from 'vscode';
+import * as vscode from "vscode";
 
 export class Store {
 
     public static get rootPath(): string {
         if (vscode.workspace.workspaceFolders) {
-            return vscode.workspace.workspaceFolders[0].uri.fsPath.replace(/\\/g, `/`);
+            return vscode.workspace.workspaceFolders[0].uri.fsPath.replace(/\\/g, "/");
         }
 
-        return ``;
+        return "";
     }
 
     public static extensionContext: vscode.ExtensionContext | undefined;
@@ -18,13 +18,13 @@ export class Store {
     public static panelViewWebviewContext: vscode.ExtensionContext | undefined;
 
     public static terminalColorsCssVars = [
-        `--vscode-terminal-ansiBlack`,
-        `--vscode-terminal-ansiRed`,
-        `--vscode-terminal-ansiGreen`,
-        `--vscode-terminal-ansiYellow`,
-        `--vscode-terminal-ansiBlue`,
-        `--vscode-terminal-ansiMagenta`,
-        `--vscode-terminal-ansiCyan`,
-        `--vscode-terminal-ansiWhite`,
+        "--vscode-terminal-ansiBlack",
+        "--vscode-terminal-ansiRed",
+        "--vscode-terminal-ansiGreen",
+        "--vscode-terminal-ansiYellow",
+        "--vscode-terminal-ansiBlue",
+        "--vscode-terminal-ansiMagenta",
+        "--vscode-terminal-ansiCyan",
+        "--vscode-terminal-ansiWhite",
     ];
 }